+++ lib/ggity/geom/line.ex
@@ -54,6 +54,9 @@ defmodule GGity.Geom.Line do
         %date_time{} when date_time in [DateTime, NaiveDateTime] ->
           Scale.X.DateTime.new(values)
 
+        value when is_binary(value) ->
+          Scale.X.Discrete.new(values)
+
         _value ->
           Scale.X.Continuous.new(values)
       end
@@ -218,15 +221,23 @@ defmodule GGity.Geom.Line do
   defp x_tick(%Geom.Line{} = geom_line, value, interval) do
     gridline_length = geom_line.width / geom_line.aspect_ratio + geom_line.area_padding * 2
     coord = geom_line.x_scale.inverse.(value)
+    %scale_type{} = geom_line.x_scale
+
+    minor_break =
+      if scale_type == Scale.X.Discrete do
+        []
+      else
+        Draw.line(
+          y2: "-#{gridline_length}",
+          stroke: "white",
+          stroke_width: "0.5",
+          transform: "translate(#{interval}, 0)"
+        )
+      end
 
     [
       Draw.line(y2: "-#{gridline_length}", stroke: "white", stroke_width: "1"),
-      Draw.line(
-        y2: "-#{gridline_length}",
-        stroke: "white",
-        stroke_width: "0.5",
-        transform: "translate(#{interval}, 0)"
-      ),
+      minor_break,
       Draw.text(Labels.format(geom_line.x_scale, value),
         fill: "gray",
         y: "9",